+++ Testing/MirrorSharpServices.cs
@@ -13,12 +13,18 @@ namespace MirrorSharp.Testing {
 
         internal IRoslynSourceTextGuard? RoslynSourceTextGuard { get; set; }
         internal IRoslynCompilationGuard? RoslynCompilationGuard { get; set; }
+        internal IConnectionSendViewer? ConnectionSendViewer { get; set; }
 
         /// <summary>Defines a <see cref="IExceptionLogger" /> called for any unhandled exception.</summary>
         public IExceptionLogger? ExceptionLogger { get; set; }
 
         internal ImmutableExtensionServices ToImmutable() {
-            return new ImmutableExtensionServices(SetOptionsFromClient, SlowUpdate, RoslynSourceTextGuard, RoslynCompilationGuard, ExceptionLogger);
+            return new ImmutableExtensionServices(
+                SetOptionsFromClient, SlowUpdate,
+                RoslynSourceTextGuard, RoslynCompilationGuard,
+                ConnectionSendViewer,
+                ExceptionLogger
+            );
         }
     }
 }