+++ WebAssets/ts/mirrorsharp.ts
@@ -21,7 +21,6 @@ export interface MirrorSharpOptions<TExtensionServerOptions = never, TSlowUpdate
     readonly serviceUrl: string;
 
     readonly selfDebugEnabled?: boolean;
-
     readonly language?: MirrorSharpLanguage;
 
     // See EditorOptions<TExtensionData>['on']. This is not DRY, but
@@ -38,8 +37,8 @@ export interface MirrorSharpOptions<TExtensionServerOptions = never, TSlowUpdate
         readonly serverError?: (message: string) => void;
     };
 
+    readonly noInitialConnection?: boolean;
     readonly initialServerOptions?: TExtensionServerOptions;
-
     readonly forCodeMirror?: CodeMirror.EditorConfiguration;
 }
 
@@ -49,6 +48,7 @@ export interface MirrorSharpInstance<TExtensionServerOptions> {
     getLanguage(): MirrorSharpLanguage;
     setLanguage(value: MirrorSharpLanguage): void;
     setServerOptions(value: TExtensionServerOptions): Promise<void>;
+    connect(): void;
     destroy(destroyOptions: { keepCodeMirror?: boolean }): void;
 }
 
@@ -57,15 +57,26 @@ export default function mirrorsharp<TExtensionServerOptions = never, TSlowUpdate
     options: MirrorSharpOptions<TExtensionServerOptions, TSlowUpdateExtensionData>
 ): MirrorSharpInstance<TExtensionServerOptions> {
     const selfDebug = options.selfDebugEnabled ? new SelfDebug() : null;
-    const connection = new Connection<TExtensionServerOptions, TSlowUpdateExtensionData>(options.serviceUrl, selfDebug);
+    const connection = new Connection<TExtensionServerOptions, TSlowUpdateExtensionData>(
+        options.serviceUrl, selfDebug, { delayedOpen: options.noInitialConnection }
+    );
     const editor = new Editor(textarea, connection, selfDebug, options);
 
+    let connectCalled = false;
     return Object.freeze({
         getCodeMirror: () => editor.getCodeMirror(),
         setText: (text: string) => editor.setText(text),
         getLanguage: () => editor.getLanguage(),
         setLanguage: (value: Language) => editor.setLanguage(value),
         setServerOptions: (value: TExtensionServerOptions) => editor.setServerOptions(value),
+        connect: () => {
+            if (!options.noInitialConnection)
+                throw new Error('Connect can only be called if options.noInitialConnection was set.');
+            if (connectCalled)
+                throw new Error('Connect can only be called once per mirrorsharp instance (on start).');
+            connection.open();
+            connectCalled = true;
+        },
 
         destroy(destroyOptions?: { keepCodeMirror?: boolean }) {
             editor.destroy(destroyOptions);